+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.0.3
 * Counter examples are automatically stored and reapplied until the properties work
   or the counter examples are deleted. See https://github.com/alfert/propcheck/pull/18
+* Mix configuration for counter examples file and for inspecting and cleaning
+  counter examples.
 
 ## 0.0.2
 * Fixed a lot of 1.5 (and 1.4) Elixir warnings thanks to https://github.com/evnu